requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Deprecated==1.2.14
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
pycparser==2.21
pydantic==2.0
pydantic_core==2.0.1
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vonage-jwt==1.0.0
wrapt==1.15.0
